chore:be(cart): create cart model
