--- backend/cart-service/src/main/java/com/nomnom/cart_service/CartItem.java
@@ -1,12 +1,0 @@
-package com.nomnom.cart_service;
-
-import lombok.Data;
-
-@Data
-public class CartItem {
-
-    private String itemId;
-    private String itemName;
-    private int quantity;
-    private double price;
-}
